+++ src/epblas-parallel/kind10/eaxpy.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 /* eaxpy — kind10 real: Y := α·X + Y. */
 typedef long double T;
 
@@ -5,7 +6,7 @@ void eaxpy_(const int *n_, const T *alpha_,
             const T *x, const int *incx_,
             T *y, const int *incy_)
 {
-    const int n = *n_, incx = *incx_, incy = *incy_;
+    const ptrdiff_t n = *n_, incx = *incx_, incy = *incy_;
     const T alpha = *alpha_;
     if (n <= 0 || alpha == 0.0L) return;
     if (incx == 1 && incy == 1) {
@@ -15,9 +16,9 @@ void eaxpy_(const int *n_, const T *alpha_,
          * epblas-openblas daxpy head) halves the per-element increment/compare/
          * branch cost vs a 4-way body and recovers ~3% over the L1-resident
          * range. Scalar tail handles n % 8. */
-        const int m = n % 8;
-        for (int i = 0; i < m; ++i) y[i] += alpha * x[i];
-        for (int i = m; i < n; i += 8) {
+        const ptrdiff_t m = n % 8;
+        for (ptrdiff_t i = 0; i < m; ++i) y[i] += alpha * x[i];
+        for (ptrdiff_t i = m; i < n; i += 8) {
             y[i    ] += alpha * x[i    ];
             y[i + 1] += alpha * x[i + 1];
             y[i + 2] += alpha * x[i + 2];
@@ -28,8 +29,8 @@ void eaxpy_(const int *n_, const T *alpha_,
             y[i + 7] += alpha * x[i + 7];
         }
     } else {
-        int ix = (incx < 0) ? (-n + 1) * incx : 0;
-        int iy = (incy < 0) ? (-n + 1) * incy : 0;
-        for (int i = 0; i < n; ++i) { y[iy] += alpha * x[ix]; ix += incx; iy += incy; }
+        ptrdiff_t ix = (incx < 0) ? (-n + 1) * incx : 0;
+        ptrdiff_t iy = (incy < 0) ? (-n + 1) * incy : 0;
+        for (ptrdiff_t i = 0; i < n; ++i) { y[iy] += alpha * x[ix]; ix += incx; iy += incy; }
     }
 }
